changes s_int_pow exponent arg. type spec to u32

+++ native/explorer/src/series.rs
@@ -709,9 +709,9 @@ pub fn s_pow(data: ExSeries, exponent: f64) -> Result<ExSeries, ExplorerError> {
 }
 
 #[rustler::nif]
-pub fn s_int_pow(data: ExSeries, exponent: i64) -> Result<ExSeries, ExplorerError> {
+pub fn s_int_pow(data: ExSeries, exponent: u32) -> Result<ExSeries, ExplorerError> {
     let s = &data.resource.0;
-    let s = s.i64()?.apply(|v| v.pow(exponent as u32)).into_series();
+    let s = s.i64()?.apply(|v| v.pow(exponent)).into_series();
     Ok(ExSeries::new(s))
 }
